Space button refreshes screen
Usability hack:  Because the system does not continously sync state with
the server, the volume and mute controls get out of sync when an
alternate client is also used.  Added the space bar to refresh the
screen state.  The usability hack works via a functional hack:  client
volume is set to the current volume.